refactor: TMS-2210 rename custom buttons
Renamed the custom button components so that they're less like vuetify names.
